made the code readable my lord what was i looking at (sorry - i dont wanna be mean)
