+++ Gostav.go
@@ -5,7 +5,6 @@ import (
 	"log"
 	"os"
 	"os/signal"
-	"strings"
 
 	"github.com/bwmarrin/discordgo"
 )
@@ -35,176 +34,8 @@ var (
 	dmPermission                   = false
 	defaultMemberPermissions int64 = discordgo.PermissionManageServer
 
-	commands = []*discordgo.ApplicationCommand{
-		{
-			Name:        "mensa",
-			Description: "Yields the menu of the Mensa INF 306.",
-		},
-
-		{
-			Name:        "impressum",
-			Description: "Reveals the src code and its coder.",
-		},
-
-		{
-			Name:        "terminal",
-			Description: "Execute the given line and yields result.",
-			Options: []*discordgo.ApplicationCommandOption{
-				{
-					Name:        "passw",
-					Description: "Password for authentification.",
-					Type:        discordgo.ApplicationCommandOptionString,
-					Required:    true,
-				},
-				{
-					Name:        "prog",
-					Description: "Program executed in the console",
-					Type:        discordgo.ApplicationCommandOptionString,
-					Required:    true,
-					// Commands might have choices, think of them like of enum values
-				},
-				{
-					Name:        "addargs",
-					Description: "Additional arguments executed after the program",
-					Type:        discordgo.ApplicationCommandOptionString,
-					Required:    false,
-					// Commands might have choices, think of them like of enum values
-				},
-			},
-		},
-		{
-			Name:        "test-command",
-			Description: "Testing message returns via functions",
-		},
-
-		{
-			Name:        "basic-command-with-files",
-			Description: "Basic command with files",
-		},
-
-		{
-			Name:        "b-field",
-			Description: "Plot the local magnetic field of my room.",
-		},
-	}
-
-	commandHandlers = map[string]func(s *discordgo.Session, i *discordgo.InteractionCreate){
-
-		"mensa": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-			s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-				Type: discordgo.InteractionResponseChannelMessageWithSource,
-				Data: &discordgo.InteractionResponseData{
-					Content: mensa_scrap(false),
-				},
-			})
-		},
-
-		"terminal": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-			options := i.ApplicationCommandData().Options
-
-			optionMap := make(map[string]*discordgo.ApplicationCommandInteractionDataOption, len(options))
-			for _, opt := range options {
-				optionMap[opt.Name] = opt
-			}
-
-			args := []string{}
-
-			if len(options) > 2 {
-				if option, ok := optionMap["passw"]; ok {
-					args = append(args, option.StringValue())
-				}
-
-				if option, ok := optionMap["prog"]; ok {
-					args = append(args, option.StringValue())
-				}
-
-				if option, ok := optionMap["addargs"]; ok {
-					args = append(args, option.StringValue())
-				}
-
-				cmds := strings.Split(args[2], " ")
-
-				s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-					Type: discordgo.InteractionResponseChannelMessageWithSource,
-					Data: &discordgo.InteractionResponseData{
-						Content: terminal(args[1], cmds, args[0]),
-					},
-				})
-			} else {
-				if option, ok := optionMap["passw"]; ok {
-					args = append(args, option.StringValue())
-				}
-
-				if option, ok := optionMap["prog"]; ok {
-					args = append(args, option.StringValue())
-				}
-
-				s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-					Type: discordgo.InteractionResponseChannelMessageWithSource,
-					Data: &discordgo.InteractionResponseData{
-						Content: terminal2(args[1], args[0]),
-					},
-				})
-			}
-
-		},
-
-		"test-command": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-			s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-				Type: discordgo.InteractionResponseChannelMessageWithSource,
-				Data: &discordgo.InteractionResponseData{
-					Content: test(),
-				},
-			})
-		},
-
-		"impressum": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-			s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-				Type: discordgo.InteractionResponseChannelMessageWithSource,
-				Data: &discordgo.InteractionResponseData{
-					Content: impressum(),
-				},
-			})
-		},
-
-		"basic-command-with-files": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-			s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-				Type: discordgo.InteractionResponseChannelMessageWithSource,
-				Data: &discordgo.InteractionResponseData{
-					Content: "Hey there! Congratulations, you just executed your first slash command with a file in the response",
-					Files: []*discordgo.File{
-						{
-							ContentType: "text/plain",
-							Name:        "test.txt",
-							Reader:      strings.NewReader("Hello Discord!!"),
-						},
-					},
-				},
-			})
-		},
-
-		"b-field": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-			dat_path := "/home/potato/Documents/sensorsave.dat"
-			img_path := b_plot(dat_path)
-			f, err := os.Open(img_path)
-			if err != nil {
-				fmt.Println(err)
-			}
-			defer f.Close()
-			s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-				Type: discordgo.InteractionResponseChannelMessageWithSource,
-				Data: &discordgo.InteractionResponseData{
-					Content: "Here's the local B-field of my room:",
-					Files: []*discordgo.File{
-						&discordgo.File{
-							Name:   img_path,
-							Reader: f,
-						},
-					},
-				},
-			})
-		},
-	}
+	commands        = all_commands
+	commandHandlers = all_handlers
 )
 
 func init() {
