+++ cmd/test.go
@@ -1,0 +1,12 @@
+package cmd
+
+import "github.com/bwmarrin/discordgo"
+
+func TestHandler(s *discordgo.Session, i *discordgo.InteractionCreate) {
+	s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+		Type: discordgo.InteractionResponseChannelMessageWithSource,
+		Data: &discordgo.InteractionResponseData{
+			Content: "Test!",
+		},
+	})
+}
