+++ go.mod
@@ -1,6 +1,6 @@
 module testgo
 
-go 1.21.0
+go 1.21
 
 require (
 	github.com/PuerkitoBio/goquery v1.8.1 // indirect
